requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
django-debug-toolbar==3.2.1
django-extra-checks==0.11.0
django-split-settings==1.0.1
django-widget-tweaks==1.4.8
psycopg2==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
